internal/cpu: Expose feature flags for ARM CPUs
This change exposes feature flags for ARM CPUs via auxv's HWCAP bits.
Because Go does not support ARMv4 and below, we do not expose
HWCAP_26BIT, since it will always be false, and do not make HWCAP_SWP
and HWCAP_HALF options for debugging, since they are mandatory for ARMv5
and up. Receiving all of the feature bits in internal/cpu required that
_AT_HWCAP2 be declared for freebsd in runtime. Additionally, hwcap and
hwcap2 are now linknamed by linux and freebsd.

Cleanup: This change enabled the declaration of hardDiv to be done in
one place, using cpu.ARM.HasIDIVA. hardDiv is a flag used by
runtime.udiv to take a fastpath if there is a hardware divider on ARM.
Also, the VFP checks in os_linux_arm.go and os_freebsd_arm.go now use
!cpu.ARM.HasVFP and !cpu.ARM.HasVFPv3, respectively.

Fixes #26630.

Change-Id: Ibbc04fb24c2b4d994f93762360f1a37bc6d83ff7