database/sql: allow Scan to work with structs
The default behavior of structs is now to sequentially scan the struct fields.
If some of the fields are unexported, the new implementation will defer to the
old behavior. This change does not affect any types that already implement the
Scanner interface.

Testing for embedded struct added.

This PR will be imported into Gerrit with the title and first
comment (this text) used to generate the subject and body of
the Gerrit change.

Updates #22697 (database/sql wishlish)
Fixes jmoiron/sqlx#162

Change-Id: I5f90b4a9deda6987306a51e708f5c4a143341a45
GitHub-Last-Rev: 409973a6ae356c95d14380d3b1e768dc608e0039
GitHub-Pull-Request: golang/go#26646