Update patch set 3
Patch Set 3: Code-Review-2

Thank you for this CL. Typically before we add a feature we create an issue "proposal: do such and such", the proposal is accepted, and then code is written.

Sometimes we make initial implementations to go along with the proposal to help see how it would look, even if the initial implementation isn't used.

This adds significant behavior to database/sql. Please open a new issue beginning with "proposal: ...", mention me "@kardianos", and we can talk about the design. In the mean time I'm marking this as won't merge. To be upfront, I would like to do this, but chances are it won't be in this manner.

Patch-set: 3
Reviewer: Gerrit User 5080 <5080@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-2