Update patch set 2
Patch Set 2:

This CL contains the fix for a bug plus a testcase. When it was merged upstream, the testcase failed on darwin, identifying a bug which was later fixed upstream. So I'm not sure if the fix for the bug uncovered by the testcase needs to be backported too, or if the backport for this fix should omit the testcase.

In any case I'd like to see the fix merged to the go 1.10 branch so it is in go 1.10.4 but don't think it can unless there is a resolution for the testcase.

Patch-set: 2
Reviewer: Gerrit User 6320 <6320@62eb7196-b449-3ce5-99f1-c037f21e1705>