Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I'm not completely convinced yet that we should omit the standard library from all. It's a weird special case.

Agreed. Looking at the bug, I think I would argue that "all" should include the subset of standard-library dependencies that are actually needed by the current module (and its transitive imports), which may be the incidental behavior from before this change.

If you agree that's the desired behavior, I'll send a change to lock it in with a test instead of changing it.

Patch-set: 2
Reviewer: Gerrit User 6365 <6365@62eb7196-b449-3ce5-99f1-c037f21e1705>