Update patch set 1
Patch Set 1:

Filippo's call really but:

I think this makes sense for crypto/x509 because there is a huge, long tail of X.509 uses and we certainly don't try to meet them all. (And nor should we, I believe.)

For the system root pool, it's a little more questionable: we have less control over the answer for one so making assertions might be making claims that we can't meet. Also, hopefully roots that are only for, say, S/MIME should have EKUs asserting that. However, if there was a root set (like Mozilla's) that only certified a subset of certificates for WebPKI use, then I do believe that it would be appropriate to return only those in crypto/x509.

Patch-set: 1
Reviewer: Gerrit User 5425 <5425@62eb7196-b449-3ce5-99f1-c037f21e1705>