cmd/go/internal/modfetch: add unit tests to downloadZip
downloadZip is in charge of looking for upstream source of a module
and downloading its zip format. It also checks if the zip format is correct.
This unit tests makes sure that downloadZip will error out if the zip format
is incorrect. CL 125436 fixed a panic that this function caused, this CL
adds the test part for it.

As part of the test, this CL also adds a mock proxy server so that future
tests can use it to create certain test scenarios that come from proxyURL.

Related issue: https://github.com/golang/go/issues/26536

Change-Id: Ic41414aacae15ff8b7d3de33a30f6ad60a5259fa
GitHub-Last-Rev: 243831ed32299fe070d93b55e68f2eb04a9d2006
GitHub-Pull-Request: golang/go#26665