Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3: Code-Review-1
> > 
> > erfc(-3.0) = 1.99997790950300147195
> > 
> > I don't think that's close enough to 2 to declare that erfc(x) = 2 for x <= -3.0.
> > That's an error of something like 1e11 epsilon.
> > The argument of erfc needs to get to around -6 before erfc(x) is within an epsilon of 2.
> 
> Speaking of which, I now understand why the Go code has a higher too-big threshold for erfc: erfc approaches zero in this case, so epsilon gets a lot smaller in that direction.

All reasonable comments. My math guru tells me that erfc will be called in practice with positive arguments where the result is near zero. That is why the function is different than erf -- it's intended to give good results near zero. But it is a problem that my implementation doesn't get closer to zero than it does for larger positive arguments. 

Meanwhile, it passes with flying colors (and gets exactly the same answers as the Go version) in the range -2.11 to 15, and runs ~2x faster than the Go version. Therefore for now, at least, I use the Go function outside of that range, and the faster function inside the range.

Patch-set: 3
Reviewer: Gerrit User 11191 <11191@62eb7196-b449-3ce5-99f1-c037f21e1705>