Update patch set 1
Patch Set 1:

This one is a low-win (at best) because nothing really uses this package, and when they do, they're going to need the tables. It's not like it's a huge package with different functionality with different users using a different subset. This package does approximately 1 thing. Or 2 things, if you could the two tables.

But in this CL, you protect both of the two tables behind the same initialization. The only real way to save memory here is to protect each table separately, but I'm not sure it's worth it. I'd probably just drop this CL and focus on other packages instead (as you have been).

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>