Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > Did you modify the ztypes file by hand or did you run the generator?
> 
> I used the generator, but left out the unrelated changes. These constants are rather old as in BSD 4.4 old, so I’m not too worried about value variance.
> 
> > 
> > If there was unexpected noise, you were likely using the wrong FreeBSD version.
> 
> What is the right FreeBSD version? That was `FreeBSD 11.2-RELEASE #0` which is fairly recent.
> 
> > 
> > Also, we need to update freebsd/arm and freebsd/386.
> 
> I don’t have either of these to run the generator on, but I’d be happy to apply the same changes if someone confirms that it’s the right thing to do. If not, I‘d consider this out of scope for this PR.

You should be able to run at least parts of the generator for freebsd/386 on freebsd/amd64. But these constants are the same for 386, arm and amd64, so I think it should be safe to manually apply the same changes for 386 and arm if you don't have those.

Patch-set: 3
Reviewer: Gerrit User 19560 <19560@62eb7196-b449-3ce5-99f1-c037f21e1705>