cmd/compile: optimize away deferred/go call to empty functions
The current implementation doesn't inline deferred or go functions.
Usually when inlining an empty function this results in a NOP which
removes the function from the go-assembly, but because defering or going
a function doesn't inline them, even empty functions are called.

Added a rule during the syntax tree stage of compilation to check
whether a function is empty before deferring it.
Added a test to codegen, and fixed test/run.go to work for windows
developers as well (check for double newline checked for \n\n and not
\r\n\r\n)

Fixes #26534

Change-Id: Ia5c1964d86fe5fc333d77225938abb611d58f211