cmd/cgo: generate correct column info
The current implementation doesn't take into consideration the fact that
cgo mangles the names of imported "C" declerations, and so the column
reported during errors is incorrect.

Added a /*line*/ directive to capture the original position of the used
C value before mangling, adjusting to decleration length.
Added a test to check the error's reported column.

Fixes #26745

Change-Id: I1b43d4096c446b4d6848eb13ce8bb9f3591fffa4