cmd/go/internal: factor out modload.QueryPackage and use in in modget
modload.Import contains a loop that looks for the module containing a package.
Because we overload Import to locate both packages and modules, that loop
contains a bunch of special-cases for modules with empty roots.

In this change, we factor out the loop into a new function (QueryPackage) and
use that directly in modget.getQuery. That restores the invariant that
the paths passed to modload.Import must be importable packages.

Perhaps more importantly, it also restores the invariant that we download the
source code for all packages in 'go get' before we attempt to build those
packages.

Fixes #26602.

Change-Id: I8bc8340c17f2df062d03ce720f4dc18b2ba406b2