crypto/x509: fix root CA extraction on macOS (no-cgo path)
The assumption that certificates without trust settings are trusted is
incorrect: certificates without any trust settings at all might still be
in the keychain (for example if they used to have some, or if they are
intermediates for offline verification). The only ones we can trust
unconditionally are the ones in the system roots store.

Moreover, the verify-cert invocation was not specifying the ssl policy,
defaulting instead to the basic one. We have no way of communicating
different usages in a CertPool, so stick to the WebPKI use-case as the
primary one for crypto/x509.

Updates #24652

Change-Id: Ife8b3d2f4026daa1223aa81fac44aeeb4f96528a