Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> This issue (go get pkg@version) came up on Slack and a few other places.
> 
> In the not-too-distant past (unless I'm mis-remembering), go get m@v only worked with modules, which clearly didn't install every package within the module.

As I understand it, `go get m@v` should only install the root package of the module, if it has one.

> Having go get work at a package level makes sense, but do we need/want it to install the package?

I don't know, but that's what the documentation says it should do today.
(I'm perfectly happy to learn that the behavior I expect is incorrect, but if so I think we need a doc update.)

> Does go get instead better fit as "ensure the code is available"?

Maybe, but that either breaks existing documentation for binaries that says “`go get $PKG` to install", or introduces a weird asymmetry between `go get $PKG@latest` and `go get $PKG`.

Patch-set: 4