Update patch set 2
Patch Set 2:

> Patch Set 2:
> 
> I don't know what this CL is about, but I want to note that at this point in the release cycle we should only be fixing critical bugs that have no workaround.  Please wait for 1.12 if at all possible.  Thanks.

Understood.

(In this case a small amount of code (added in a prior change) code was replaced with dedicated function calls (that have existed for years) which are strictly more safe (ie, robust in the presence of unexpected errors) and where written explicitly for the task that the "hand-written" code was trying to achieve in the first place. Also, this code is only executed in the presence of (type-checking) errors. The change is very safe, but I agree that the rationale for submitting this so soon before the first RC is not strong.)

Patch-set: 2