Update patch set 19
Patch Set 19: Run-TryBot+1

(1 comment)

> Patch Set 19:
> 
> (2 comments)
> 
> Sorry for replying late. There is a trivial issue with current frame design. We will need special handling when calling to native function directly without going through cgo. I worked around this by reserving additional 16 bytes.

Seems ok.

> 
> It won't cause problems if the developers won't call native function directly with Go assembly code. But I am not sure if this is allowed or not.

Calling external function directly from user assembly code is not supported. We don't need to worry about it. It might be possible to carefully code it to get it work for certain special cases, but we never guaranteed, or even tried, to make it work. Note that in the few special cases in the runtime, the external function are all called on special stacks (system stack or signal stack), which is already not the user code supposed to do.

Patch-set: 19
Reviewer: Gerrit User 13315 <13315@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Run-TryBot=+1