Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> (1 comment)
> 
> I've made the changes, but I'm curious how we want developers to discover this functionality, given that the constants are now unexported. Is the issue going to be the only way for users to see this?

It's just too late in the cycle to add API.

But as Wasm is experimental, we could experiment with this behavior, without public docs & API for now. It can be documented on a wiki page or other forums for now until we decide whether we're happy with it.

Patch-set: 4
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>