Update patch set 1
Patch Set 1:

I might be wrong, but the way the tests for the new functions are written (in the other CL) makes it so these rules don’t trigger while running them. In other words, it seems to me that these rules are actually untested in behavior; the codegen tests just make sure they trigger (and they are useful) but they could produce wrong code and nobody would notice. I think we need a way to make sure that a correctness testsuite is also run for the intrinsified version.

Patch-set: 1
Reviewer: Gerrit User 5340 <5340@62eb7196-b449-3ce5-99f1-c037f21e1705>