Update patch set 1
Patch Set 1: Code-Review-1

I think it was better before. Before it said "and so on", so it was never inaccurate.

Now it pretends to be authoritative, so once it's out of date, the docs are wrong.

To get the complete list, run "go tool dist list".

Patch-set: 1
Reviewer: Gerrit User 5065 <5065@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=-1