cmd/go: ignore empty gofiles when running list
go list shouldn't return a non-zero exit status if a package
only contains test files. This change will ignore empty gofiles
when running a build action for a go list.

I don't know if this is the correct way to fix the issue but
I've included a test that exercises the bug.

Fixes #27097

Change-Id: I3ccc509adbfb76753d54b959716ae94a4bf30a9f