Update patch set 17
Patch Set 17:

> What about OpenBSD and DragonFly?
> 
> I think we should probably just say that the answer here is to use x/sys/unix.
> 
> Tobias, opinion?

I'm inclined to think that since the constants are defined in the syscall package on Darwin, GNU/Linux, and NetBSD that it's OK to define them in the syscall package on other systems as well.  That will promote portability for packages that currently use syscall, even if it would be better in the long run for them to move to x/sys/unix.

Patch-set: 17
CC: Gerrit User 5206 <5206@62eb7196-b449-3ce5-99f1-c037f21e1705>