Update patch set 1
Patch Set 1:

(1 comment)

> Patch Set 1:
> 
> (2 comments)
> 
> Can you update the CL description to show the new output?  Thanks.

The title has changed to a this peculiar flag but I opened
it initially because the information about the default flags passed
to vet was difficult to find.

With this information and GOFLAGS, it is easy to understand and to
configure the actions of go vet. I'll set updates iso fixes.

Patch-set: 1
Reviewer: Gerrit User 26957 <26957@62eb7196-b449-3ce5-99f1-c037f21e1705>