Update patch set 1
Patch Set 1:

(2 comments)

This seems ok in principle, but I have some questions re: over-/underflow of the exponent. Do we know that it cannot happen? (If so, it needs comments). If it can happen, we need to handle it.

Patch-set: 1
Reviewer: Gerrit User 5210 <5210@62eb7196-b449-3ce5-99f1-c037f21e1705>