Update patch set 2
Patch Set 2:

(2 comments)

Thanks for doing this. I think this is a very reasonable fix. Constants above 2**63 used in such a context are rare and walkinrange punts on many other hard cases.

This says "DO NOT REVIEW", but I came in via the linked issue; feel free to completely ignore my comments.

Patch-set: 2
CC: Gerrit User 5143 <5143@62eb7196-b449-3ce5-99f1-c037f21e1705>