Update patch set 14
Patch Set 14:

> I don't know what you mean by "propagates an empty network" and I see no ambiguity.

hm, can you please explain what does an empty network mean? the exposed api surface rejects an empty network to avoid such ambiguity.

for example, this change introduces the following:

> addrs, err, completed := cgoLookupIP(ctx, "", name)

in context (that comes from a function name in this example), it can assume that "ah, perhaps part of ip protocol suite", but the parameter network is used anywhere, not only lookup code path, because that's a primary hint for the succeeding address parameter. 

i still don't see any good reason to use "" instead of "ip" for the scope of ip protocol suite, "unspec", "unspecified" or "this-is-an-unspecified-network".

> But I'm making the call here: we keep the function.

then, the func should have a more specific name and move to dial.go (just after parseNetwork) or ipsock.go because it interprets the network parameter in some scope.

Patch-set: 14
Reviewer: Gerrit User 5076 <5076@62eb7196-b449-3ce5-99f1-c037f21e1705>