Update patch set 7
Patch Set 7: Code-Review+2

> Patch Set 7:
> 
> (1 comment)

I see now.  With SUBQload the value in memory is the subtrahend.  For SUBQmodify the value in memory is the minuend.  For symmetric ops the rule is redundant, but not for the non-symmetric ones.

So this looks ok.

Patch-set: 7
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>
Label: Code-Review=+2