Update patch set 2
Patch Set 2:

And just one final thought, since AuxInt defaults to 0, 0 should really mean "might be divide by -1".  1 should mean "known to not be -1". That way that bit is set correctly when the value is constructed, so if, for example, the prove pass is skipped the code is still correct.

Patch-set: 2
Reviewer: Gerrit User 5200 <5200@62eb7196-b449-3ce5-99f1-c037f21e1705>