Update patch set 1
Patch Set 1:

Polyfills should be provided by the user. Please move this to the other polyfill in wasm_exec.html.

Who is the author of this polyfill? Are we allowed to use it? Do we need to add attribution?

Patch-set: 1
Reviewer: Gerrit User 13620 <13620@62eb7196-b449-3ce5-99f1-c037f21e1705>