Update patch set 2
Patch Set 2:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > Why are you reverting on the branch? Is there an immediate release being cut? If not, it would be better to wait for the fix and evaluate if the fix can be backported.
> > 
> > It's a pretty bad bug, and I couldn't find the error reading the CL, nor could Josh (we're both at Gophercon).  I thought we should have a revert in hand at least, so we have the option of a very quick fix.
> > 
> > I'll look at the CL some more to try figure out what went wrong and devise a better fix, or if you understand what went wrong, even better.
> 
> I agree the bug is bad, prove is really delicate in that it drops a basic security mitigation and it's hard to properly test it. I don't want to sound conceited but I don't think it'd take me long to find the bug given that I wrote it. The issue is, I'm in the middle of packing and I have a flight to catch. I can have a look during the flight.
> 
> Up to you if going on with the reversion, but if there's no incoming release maybe you can wait a couple of days and we can see how invasive is the fix and how broken was the CL.

I am happy to wait on your plane flight. You do not sound conceited; provers are tricky.  This does argue for better documentation, however, if we are dependent on someone with the author's understanding to fix it quickly/confidently.  I imagine there is a missing "x must be >= k" test against the existing constraints somewhere, but I don't know where to put it or exactly how to say it (but I will look).

The biggest reason for the revert is so that we have it there, and ready, to help bound nervousness if someone decides it really must be fixed immediately.

Patch-set: 2
Reviewer: Gerrit User 7061 <7061@62eb7196-b449-3ce5-99f1-c037f21e1705>